+++ public/browser/chunk-6jCDkiiK.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-DkxHLRVR.js`).then(o=>o.LeadsListComponent)},{path:`:id`,loadComponent:()=>import(`./chunk-CsZmYjRq.js`).then(o=>o.LeadDetailComponent)}];export{t as LEADS_ROUTES};