+++ public/browser/chunk-BREEDh9T2.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";import"./chunk-B_m5Yd57.js";import"./chunk-C53ZhtzR.js";import"./chunk-BRzOHBVt.js";import"./chunk-BBc5x1WO.js";import"./chunk-BmpWjZKS.js";import"./chunk-BgjWCWRg.js";import"./chunk-D6Cr68VN.js";import"./chunk-D3kTOyaV.js";import{P as Fe}from"./main-SCJJN7TP.js";export{Fe as BpmnEditorPage};