+++ public/browser/chunk-DibFFCnG.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-DkK6UCG-2.js`).then(o=>o.DefinitionsListPageComponent)}];export{t as DEFINITION_ROUTES};