+++ public/browser/chunk-wLJ86KGZ.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-B28kIYs82.js`).then(o=>o.InboxListComponent)}];export{t as INBOX_ROUTES};